requirements.txt
whisperx>=3.8
python-dotenv>=1.0
